Main
Here's delve clicker. This is cmd game for now. It's first version of game that i publishing. Later, when all options will be finished i'll port this to GUI interface. But now this is first 0.1 version. You may play it on your cmd if you launch it. If you will find any bug, fix it, or just tell about it to me. I will be grateful to you. Big Thank You!